requirements.txt
discord.py; python_version >= "3.13"